refactor Org settings screen and add change treasury btn
