+++ lib/screens/organization/org_settings/api.dart
@@ -1,0 +1,24 @@
+import 'package:iw_app/api/models/org_offers_filter_model.dart';
+import 'package:iw_app/api/orgs_api.dart';
+import 'package:iw_app/models/offer_model.dart';
+import 'package:iw_app/models/organization_member_model.dart';
+import 'package:iw_app/models/organization_model.dart';
+
+Future<List<Offer>> fetchInvestOffers(Organization organization) async {
+  try {
+    final filter = OrgOffersFilter(
+      status: OfferStatus.Pending,
+      role: MemberRole.Investor,
+    );
+    final response = await orgsApi.getOffers(
+      organization.id!,
+      filter,
+    );
+    final offers =
+        (response.data as List).map((json) => Offer.fromJson(json)).toList();
+    return offers;
+  } catch (err) {
+    print(err);
+  }
+  return [];
+}
